Updated property type

+++ Bex/Model/ActivitySegment.cs
@@ -22,7 +22,7 @@ namespace Bex.Model
         public int SplitDistance { get; set; }
 
         [JsonProperty("segmentId")]
-        public int SegmentId { get; set; }
+        public long SegmentId { get; set; }
 
         [JsonProperty("startTime")]
         public DateTime? StartTime { get; set; }
